Move bpfd-agent and bpfd-operator code
Change code location to `controllers/` to re-align
with operator SDK best practices. This also opens
up `pkg` so we can write a library for application
users.
