+++ bpfd-operator/Makefile
@@ -152,8 +152,8 @@ endif
 .PHONY: manifests
 manifests: controller-gen ## Generate WebhookConfiguration, ClusterRole and CustomResourceDefinition objects.
 	$(CONTROLLER_GEN) crd webhook paths="./..." output:crd:artifacts:config=config/crd/bases
-	$(CONTROLLER_GEN) rbac:roleName=agent-role paths="./pkg/bpfd-agent" output:rbac:artifacts:config=config/rbac/bpfd-agent-role
-	$(CONTROLLER_GEN) rbac:roleName=operator-role paths="./pkg/bpfd-operator" output:rbac:artifacts:config=config/rbac/bpfd-operator-role
+	$(CONTROLLER_GEN) rbac:roleName=agent-role paths="./controllers/bpfd-agent" output:rbac:artifacts:config=config/rbac/bpfd-agent-role
+	$(CONTROLLER_GEN) rbac:roleName=operator-role paths="./controllers/bpfd-operator" output:rbac:artifacts:config=config/rbac/bpfd-operator-role
 
 
 .PHONY: generate
